update modals file
