+++ app/v2/businesses/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-businesses = Blueprint('businessesv2' , __name__)
+businesses = Blueprint('businessesv2' , __name__, url_prefix='/api/v2')
 
 from . import views
